refactor: divide ponder schema code into relations and tables files
